Progress bar (#18)
Improved function for updating the progres bar when using adaptive timing. Is much smoother now, and tpyically only jumps the last ~20% or so. Starts at a reasonable place as well.

Removed warning label at top of main page as it is no longer needed

Updated about page to include a note that we are using Drude approximations for our metals